Create Cinema.java
A simple program to manage bookings of a cinema hall room. It can print the layout of the room, with the empty and bought seats. It also manages the revenue earned and shows various statistics like the number of tickets bought, tickets bought as a percentage of total seats, and incomes.